git: Enable rerere
Reuse recorded resolution of conflicted merges.

I'm unsure that I want to automatically record and repeat the same merge
conflict resolution, because I don't feel confident that will always do
what I want. Perhaps I made a mistake the first time, and want to do it
differently the second.

It appears from the man page that recorded resolutions will be cleared
when a rebase or patch application is aborted:

    clear
        Reset the metadata used by rerere if a merge resolution is to be
        aborted. Calling git am [--skip|--abort] or git rebase
        [--skip|--abort] will automatically invoke this command.