vim-plugin: Add plugin to operate on visual selections
NOTE: This plugin does not include 'doc/tags' in its gitignore, so to
prevent annoying __git_ps1 and git-status notifications of untracked
content, do

    echo "doc/tags" >> .git/modules/vim/bundle/vis/info/exclude

Ex commands on visual ranges are by default linewise. That is, even if
the visual selection is a fragment of a line, or a visual block, running

    :'<,'> !cat

will print the whole line(s) selected.

vis.vim allows running Ex commands on _just_ the selection region with

    :'<,'>B !cat

For example, take a log file with Unix timestamps

    (1630780361.234764) log stuff here
    (1630780463.123763) log stuff here
    (1630780541.895731) log stuff here

Use visual block mode to select the columns inside the parenthesis, and
run

    :'<,'>B !xargs -I\% date \d @\% --utc --iso-8601=ns<CR>

to convert the timestamps to UTC ISO-8601 timestamps.

    (2021-09-04T18:32:41,234764000+00:00 ) log stuff here
    (2021-09-04T18:34:23,123763000+00:00 ) log stuff here
    (2021-09-04T18:35:41,895731000+00:00 ) log stuff here

The equivalent shell one-liner would be

    sed 's/(\(.*\))\(.*\)/printf "%s" "("; date --utc --iso-8601=ns -d @\1 | tr -d "\n"; printf ")%s" "\2"/e'
                          ================================================================================

where the underscored portion of the

    s/pattern/replacement/e
              ===========

is a shell command (or sequence of commands) that's evaluated with the
values of the various capture groups.

    $ cat test.txt | sed 's/(\(.*\))\(.*\)/printf "%s" "("; date --utc --iso-8601=ns -d @\1 | tr -d "\n"; printf ")%s" "\2"/e'
    (2021-09-04T18:32:41,234764000+00:00) log stuff here
    (2021-09-04T18:34:23,123763000+00:00) log stuff here
    (2021-09-04T18:35:41,895731000+00:00) log stuff here

Notice that the commands must printf the parts of the line that _aren't_
being transformed. This is because simply using

    $ cat test.txt | sed 's/(\(.*\))\(.*\)/date --utc --iso-8601=ns -d @\1/e'
    2021-09-04T18:32:41,234764000+00:00
    2021-09-04T18:34:23,123763000+00:00
    2021-09-04T18:35:41,895731000+00:00

prints just the output of the date command.