config: Prefer unofficial Teams client
Microsoft is going to deprecate the native Linux client in favor of
progressive web apps, so switch _back_ to the unofficial teams client,
which apparently  has worked better for a while now anyways (screen
sharing on Wayland and being able to see more than four people in a
call).

Installing the unofficial teams-for-linux client from snap 1.0.29 on the
stable channel didn't work when attempting to screenshare on Linux, so
go back to the latest releast from GitHub.