vim: Switch to stable rust-analyzer
The motivation for this change was the enable "cargo build" and
rust-analyzer to share the build cache, which should hopefully result in
fewer rebuilds (time and space!!).

However, I've found that using the stable rust-analyzer (which should
invoke the stable toolchain when building) had no effect. All of my
googling told me that it's likely a difference in environment variables
between the r-a env and the cargo env, but I set RUSTFLAGS and all of
the same environment variables that ./bashrc.d/07-environment.sh sets,
and it still rebuilt everything.