WIP: bashrc: Enable hcmnt
TODO: This runs stuff in the background I think, which causes a [1] in
my PS1.
TODO: A comment on the post said that in order to use with history -a,
you have to put the history log in another file.
TODO: The ~/.history_log doesn't have the commands in it - only
timestamps and the CWD.
TODO: The timestamps don't have high enough granularity